// sanity.types.ts
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type LocalizedString = {
  _type: "localizedString";
  en?: string;
  ro?: string;
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: sanity/lib/queries.ts
// Variable: PROJECTS_QUERY
// Query: *[_type == "project"] | order(order asc) {    _id, title, slug, coverImage, category->{title, slug}, location, featured  }
export type PROJECTS_QUERY_RESULT = Array<{
  _id: string;
  title: LocalizedString | null;
  slug: Slug | null;
  coverImage: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
  category: {
    title: LocalizedString | null;
    slug: Slug | null;
  } | null;
  location: string | null;
  featured: boolean | null;
}>;